+++ issue10/src/index.jsx
@@ -1,5 +1,5 @@
 import {createRoot} from "react-dom/client";
-import {App}  from "./App";
+import {App} from "./App";
 
 const container = document.getElementById("app");
 const root = createRoot(container)
